Rename iconPicker to getIcon
